refactor : 패키지 구조 수정
1. User관련 패키지 구조 수정
