--- src/main/java/com/flab/shoeauction/domain/user/UserLevel.java
@@ -1,5 +1,0 @@
-package com.flab.shoeauction.domain.user;
-
-public enum UserLevel {
-    USER, ADMIN
-}
