+++ src/main/java/com/flab/shoeauction/domain/users/admin/AdminRepository.java
@@ -1,6 +1,5 @@
-package com.flab.shoeauction.domain.user.repository;
+package com.flab.shoeauction.domain.users.admin;
 
-import com.flab.shoeauction.domain.user.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminRepository extends JpaRepository<Admin, Long> {
