+++ src/main/java/com/flab/shoeauction/domain/users/common/UserLevel.java
@@ -1,0 +1,5 @@
+package com.flab.shoeauction.domain.users.common;
+
+public enum UserLevel {
+    USER, ADMIN
+}
