+++ src/main/java/com/flab/shoeauction/service/SessionLoginService.java
@@ -5,8 +5,8 @@ import static com.flab.shoeauction.common.utils.user.UserConstants.USER_ID;
 
 import com.flab.shoeauction.controller.dto.UserDto.LoginRequest;
 import com.flab.shoeauction.controller.dto.UserDto.UserInfoDto;
-import com.flab.shoeauction.domain.user.User;
-import com.flab.shoeauction.domain.user.repository.UserRepository;
+import com.flab.shoeauction.domain.users.user.User;
+import com.flab.shoeauction.domain.users.user.UserRepository;
 import com.flab.shoeauction.exception.user.UserNotFoundException;
 import com.flab.shoeauction.service.encrytion.EncryptionService;
 import javax.servlet.http.HttpSession;
