+++ src/test/java/com/flab/shoeauction/service/LoginServiceTest.java
@@ -8,8 +8,8 @@ import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.flab.shoeauction.domain.user.User;
-import com.flab.shoeauction.domain.user.repository.UserRepository;
+import com.flab.shoeauction.domain.users.user.User;
+import com.flab.shoeauction.domain.users.user.UserRepository;
 import com.flab.shoeauction.exception.user.UserNotFoundException;
 import com.flab.shoeauction.service.encrytion.EncryptionService;
 import com.flab.shoeauction.controller.dto.UserDto.LoginRequest;
