+++ src/test/java/com/flab/shoeauction/service/UserServiceTest.java
@@ -16,9 +16,9 @@ import com.flab.shoeauction.controller.dto.UserDto.SaveRequest;
 import com.flab.shoeauction.domain.addressBook.Address;
 import com.flab.shoeauction.domain.addressBook.AddressBook;
 import com.flab.shoeauction.domain.addressBook.AddressBookRepository;
-import com.flab.shoeauction.domain.user.Account;
-import com.flab.shoeauction.domain.user.User;
-import com.flab.shoeauction.domain.user.repository.UserRepository;
+import com.flab.shoeauction.domain.users.common.Account;
+import com.flab.shoeauction.domain.users.user.User;
+import com.flab.shoeauction.domain.users.user.UserRepository;
 import com.flab.shoeauction.exception.user.DuplicateEmailException;
 import com.flab.shoeauction.exception.user.DuplicateNicknameException;
 import com.flab.shoeauction.exception.user.UnableToChangeNicknameException;
